// Backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PassStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  NONE: 'NONE'
} as const

export type PassStatus = (typeof PassStatus)[keyof typeof PassStatus]


export const userType = {
  ADMIN: 'ADMIN',
  USER: 'USER',
  SUBADMIN: 'SUBADMIN'
} as const

export type userType = (typeof userType)[keyof typeof userType]


export const EventType = {
  TECH: 'TECH',
  NON_TECH: 'NON_TECH',
  PASS: 'PASS'
} as const

export type EventType = (typeof EventType)[keyof typeof EventType]